backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["sumit pandey <sumitpandey0304@gmail.com>"]
readme = "README.md"
packages = [{include = "app", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.15"
google-ads = "^28.4.1"
flask = "^3.1.2"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.5"
flask-cors = "^5.0.0"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.2.1"
marshmallow = "^3.20.1"
flask-marshmallow = "^1.2.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
